Usecase of UAVs simulated in CoppeliaSim.
Running uav.py, it will run mn-wifi and a CoppeliaSim with a
simulation of 3 Drones (simulation.ttt).

With the script simpleTest.py the Drone's position is
read from the simulation and saved in data/.

In parallel, setNodePosition.py sets the position of the
stations in mn-wifi.

For CoppeliaSim are necessary the following libraries:

  - remoteApi.so
  - sim.py
  - simConst.py