change network={} logic for wpa_sup conf given conf parameter from addStation()
By default the logic includes (or expects) some values (ssid, ieee80211w, key_mgmt) to construct the wpa_supplicant.conf with the network={}.

This change considers if the config attribute from addStation is set to "", to do not include network={}. 
If config is set as expected, it keeps working as before. If config is not set, keeps working as before.